requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Migrations
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic[email]
bleach==6.2.0

# Caching
redis==5.2.1

# Background Tasks
celery==5.4.0

# Observability
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
factory-boy==3.3.1
